Prevent client expiration while recalling a delegation
We hold the refcount on the client to make sure that it exists
throughout the handling of recall and revoke. This doesn't ensure that
the client doesn't get expired while doing this. We need to reserve the
lease to avoid the client getting expired to make sure that we can
safely delete the delegation state without racing with the reaper thread
that handles client expirations.

Signed-off-by: Malahal Naineni <malahal@us.ibm.com>
Reviewed-by: Srikrishan Malik <srimalik@in.ibm.com>