Relax permission checking on OPEN CREATE
If we have an OPEN CREATE, the file may be created with a mode that would
preclude the access type of the open. That is not what is expected, so
make a best effort determination that we created the file, and use that
to skip permission checking on open.

Note that we REALLY need to make a better effort at atomicity of OPEN CREATE.
This code is subject to a race where our lookup fails, but simultaneously
someone else creates the file in a way that would preclude this OPEN from
opening the file. Particularly insidious would be another user creating the
file...

Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>
Acked-by: Malahal Naineni <malahal@us.ibm.com>