Simplify getting clientid_owner for LAYOUT ops (fixes CID24720)
The LAYOUT ops used to call get_clientid_owner to get the clientid owner
for the clientid associated with the current session. Since the ops MUST
be within a session, and thus the session and it's associated clientid
MUST be valid, we don't need to lookup, we can just look into the
associated clientid for it's owner.

This resolves any reference counting issues, and happens to clear a
Coverity bug with an ignored return result.

Signed-off-by: Frank S. Filz <ffilzlnx@mindspring.com>
Acked-by: Malahal Naineni <malahal@us.ibm.com>