NFS4 OPEN OWNERS: Provided support to cache open owners for the lease period time instead of deleting them on close.
In the current NFSv4 implementation, open_owners are freed once
the server receives the last CLOSE for the corresponding files, i.e,
when their refcount becomes zero. This behavior differs from that of
knfs server, where in the latter server caches those owners for certain
amount of time (mostly for the lease period time), so that the client trying
to reopen that file within that period, need not perform OPEN_CONFIRM again.

Have added support to achieve the same behavior using ganesha server
as well, but with a configurable option to enable it. So with this
option set, instead of deleting the open_owners on last CLOSE op
received, we store the close timestamp in the new field added 'last_close_time'
field in the state_owner structure.
The reaper thread, which cleans up the expired clients, will now also
look up the 'ht_nfs4_owner' hashtable to check and free the entries
with refcount=0 and the (last_close_time + lease_period > current_time)

And ifincase , clients reopen the same file within the lease_period time,
the 'last_close_time' field will be reset.

Signed-off-by: Soumya Koduri <skoduri@redhat.com>