FSAL_GLUSTER: restore creds from the saved open_fd
Not all the time op_ctx shall be set, for eg., while revoking
state if the clientid is expired. Hence read the creds/lease_id
from the fd used to open the file

This change shall also be useful to close files incase
if ganesha process is started by non-root user.

Change-Id: I6aaeea504cde2fb30285ef843591a759483c6592
Signed-off-by: Soumya Koduri <skoduri@redhat.com>
Signed-off-by: Jiffin Tony Thottan <jthottan@redhat.com>